DFTDP # 282 Move submission-history to shared library - Added Submission History component to shared portal UI - Exported the componnet in lic api.ts - Updated driver portal submission history componnet to use the shared component
