+++ driver-portal/src/ClientApp/src/app/submission-history/submission-history.component.html
@@ -1,174 +1,1 @@
-<div class="container">
-  <div class="list">
-    <h2>Submission History</h2>
-    <div class="submissions">
-      <div>
-        <p>
-          The Submission History tab displays the documents and reports that
-          RSBC has received to date on all your cases. Unfortunately, you cannot
-          download a copy of these submissions.
-        </p>
-        <p>
-          Please note that if your medical practitioner or you have sent a
-          document through fax, it might take a while for it to appear as
-          recevied and shown in the list below.
-        </p>
-        <p>
-          Find out more about how to resubmit a document as well as if and how
-          to request a copy of a submission.
-        </p>
-      </div>
-      <div>
-        <app-quick-links></app-quick-links>
-      </div>
-    </div>
-  </div>
-
-  <div class="list">
-    <div>
-      <mat-card
-        *ngFor="let document of filteredDocuments"
-        [ngClass]="{
-          open: document.documentId && isExpanded[document.documentId]
-        }"
-      >
-        <mat-card-content class="closedCaseContent">
-          <div style="text-align: center">
-            <mat-icon
-              (click)="toggleIsExpandable(document.documentId)"
-              class="expandedIcon"
-            >
-              expand_circle_down
-            </mat-icon>
-          </div>
-          <div style="display: flex; flex-direction: column">
-            <div class="case-details-border">&nbsp;</div>
-            <div>&nbsp;</div>
-          </div>
-          <div style="flex: 1">
-            <div class="caseDetailRow">
-              <div>
-                <label>Created Date </label>
-                <div>
-                  <b>{{ document.createDate | date : "longDate" }}</b>
-                </div>
-              </div>
-
-              <div>
-                <label>Submission Type</label>
-                <div>
-                  <b
-                    >{{ document.documentType }} 
-                    @if(document?.documentSubType){
-                    - {{ document?.documentSubType }}
-                    }
-                  </b>
-                </div>
-              </div>
-              <div>
-                <label>Submission status</label>
-                <div>
-                  <b>{{ document.submittalStatus }}</b>
-                </div>
-              </div>
-              <div>
-                <label>Submission date</label>
-                <div>
-                  <b>{{ document.importDate | date : "longDate" }}</b>
-                </div>
-              </div>
-            </div>
-
-            <div
-              class="caseDetailRow"
-              *ngIf="document.documentId && isExpanded?.[document.documentId]"
-              style="
-                border-top: 2px solid gray;
-                padding-top: 20px;
-                padding-bottom: 20px;
-              "
-            >
-              <div>
-                <label>Description</label>
-                <div>
-                  <b>{{ document.description }}</b>
-                </div>
-              </div>
-            </div>
-          </div>
-        </mat-card-content>
-      </mat-card>
-    </div>
-    <div *ngIf ="!isLoading && submissionHistoryDocuments?.length === 0" >
-      <mat-card>
-        <mat-card-content>
-          <b>No Submission History records found for this case</b>
-        </mat-card-content>
-      </mat-card>
-    </div>
-  </div>
-
-  <div
-    class="numberOfItems"
-    *ngIf="
-    submissionHistoryDocuments &&
-    submissionHistoryDocuments.length > 0 &&
-      filteredDocuments &&
-      filteredDocuments.length > 0
-    "
-  >
-    Showing {{ filteredDocuments.length }} of
-    {{ submissionHistoryDocuments.length }} records
-    <br />
-    <a
-      *ngIf="filteredDocuments.length < submissionHistoryDocuments.length"
-      (click)="viewMore()"
-      class="viewMore"
-      ><b style="color: #003366">View more records</b>
-      <mat-icon class="ViewMoreIcon">expand_circle_down</mat-icon></a
-    >
-  </div>
-
-  <div class="list" id="definitions">
-    <h2>Definitions</h2>
-    <div class="accordion-btn-container">
-      <a class="accordion-btn show-btn" (click)="accordion.openAll()"
-        >Expand All</a
-      >
-      <span class="accordion-bar">&nbsp;|&nbsp;</span>
-      <a class="accordion-btn hide-btn" (click)="accordion.closeAll()"
-        >Collapse All</a
-      >
-    </div>
-    <div style="padding-bottom: 30px">
-      <mat-accordion multi>
-        <mat-expansion-panel>
-          <mat-expansion-panel-header class="panel-header">
-            <mat-panel-title class="accordian-Title">
-              Case Type
-            </mat-panel-title>
-          </mat-expansion-panel-header>
-          <app-case-type></app-case-type>
-        </mat-expansion-panel>
-
-        <mat-expansion-panel>
-          <mat-expansion-panel-header class="panel-header">
-            <mat-panel-title class="accordian-Title">
-              Submission Type</mat-panel-title
-            >
-          </mat-expansion-panel-header>
-          <app-submission-type></app-submission-type>
-        </mat-expansion-panel>
-
-        <mat-expansion-panel>
-          <mat-expansion-panel-header class="panel-header">
-            <mat-panel-title class="accordian-Title">
-              Submission Status</mat-panel-title
-            >
-          </mat-expansion-panel-header>
-          <app-submission-status></app-submission-status>
-        </mat-expansion-panel>
-      </mat-accordion>
-    </div>
-  </div>
-</div>
+<app-shared-submission-history [caseManagementService]="caseManagementService"></app-shared-submission-history>
