+++ shared-portal-ui/projects/core-ui/src/lib/submission-history/submission-history.component.scss
@@ -1,0 +1,80 @@
+.submissions {
+  display: flex;
+  flex-direction: row;
+  padding-top: 24px;
+  gap: 100px;
+}
+.submissionRequirement {
+  display: flex;
+  width: 100%;
+}
+
+.closedCaseContent {
+  display: flex;
+  flex-direction: row;
+  gap: 24px;
+}
+
+.case-details-border {
+  border-left: 2px solid #808080;
+  height: 45%;
+}
+
+.caseDetailRow {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  line-height: 32px;
+  padding-bottom: 24px;
+  align-content: flex-start;
+}
+
+.caseDetailRow > div {
+  flex: 1 1 220px;
+}
+
+.mat-mdc-card {
+  border: 1px solid gray;
+  box-shadow: none !important;
+  margin: 30px 0 30px 0;
+  padding: 24px;
+  .mat-mdc-card-header {
+    padding-top: 12px;
+
+    .mat-mdc-card-header-text {
+      width: 100%;
+      vertical-align: middle;
+    }
+  }
+}
+
+.expandedIcon {
+  transform: scale(1.5);
+  color: #003366;
+  width: 50px;
+}
+
+.open {
+  background-color: #f1f1f3;
+  border: 1px solid #003366;
+  cursor: pointer;
+}
+
+@media only screen and (max-width: 896px) {
+  .submissions {
+    display: block;
+  }
+  .closedCaseContent {
+    flex-direction: column-reverse;
+  }
+
+  .caseDetailRow {
+    gap: 80px;
+  }
+
+  .case-details-border {
+    border-left: none;
+    border-top: 4px solid #808080;
+    height: 8px;
+  }
+}
