+++ shared-portal-ui/projects/core-ui/src/lib/submission-history/submission-history.component.spec.ts
@@ -1,0 +1,17 @@
+describe('SubmissionHistoryComponent', () => {
+  // let component: SubmissionHistoryComponent;
+  // let fixture: ComponentFixture<SubmissionHistoryComponent>;
+
+  // beforeEach(() => {
+  //   TestBed.configureTestingModule({
+  //     declarations: [SubmissionHistoryComponent]
+  //   });
+  //   fixture = TestBed.createComponent(SubmissionHistoryComponent);
+  //   component = fixture.componentInstance;
+  //   fixture.detectChanges();
+  // });
+
+  it('should create', () => {
+    //expect(component).toBeTruthy();
+  });
+});
